requirements.txt
fastapi
uvicorn
langgraph
pydantic
python-dotenv
openai
pymongo
requests
finnhub-python
websockets>=13.0
urllib3>=2.2.2
fpdf2